Phylo package, new options for multiFaAcceleration
