+++ cmd/multiFaAcceleration/multiFaAcceleration.go
@@ -10,9 +10,8 @@ import (
 	"github.com/vertgenlab/gonomics/exception"
 	"github.com/vertgenlab/gonomics/fasta"
 	"github.com/vertgenlab/gonomics/fileio"
-	"github.com/vertgenlab/gonomics/numbers"
+	"github.com/vertgenlab/gonomics/phylo"
 	"log"
-	"math"
 )
 
 //Settings contains all program arguments and options, and is passed throughout helper functions to aid readability.
@@ -30,6 +29,8 @@ type Settings struct {
 	Epsilon                    float64
 	AllowNegative              bool
 	ZeroDistanceWeightConstant float64
+	B1Out	string
+	B3Out	string
 }
 
 //Once we have branch lengths for each valid window, we will need to normalize the values relative to each other. Thus, we store the branch lengths in this intermediate cache before writing to file.
@@ -38,36 +39,6 @@ type BranchCache struct {
 	ChromEnd   int
 	B1         float64
 	B3         float64
-}
-
-//A set of all observed pairwise distances between the four species. While these numbers must be integers, we store them as float64 to avoid casting in other functions as a way to improve readability.
-type Distances struct {
-	D01 float64
-	D02 float64
-	D03 float64
-	D12 float64
-	D13 float64
-	D23 float64
-}
-
-//The set of branch lengths corresponding to a particular distance matrix.
-type BranchLengths struct {
-	B1 float64
-	B2 float64
-	B3 float64
-	B4 float64
-	B5 float64
-}
-
-//SubTree is a tree with three leaves and three branches, joined at the single internal node.
-//Dij represent the observed pairwise distances between two species. vi represents the length of the branch between a species i and the internal node at the current stage of optimization.
-type SubTree struct {
-	Dab float64
-	Dac float64
-	Dbc float64
-	Va  float64
-	Vb  float64
-	Vc  float64
 }
 
 func multiFaAcceleration(s Settings) {
@@ -100,8 +71,8 @@ func multiFaAcceleration(s Settings) {
 		threshold = int(s.SearchSpaceProportion * float64(s.WindowSize)) //the minimum number of bases at which a window must overlap the search space in order to be considered a valid window.
 	}
 
-	var currDistances Distances
-	var distanceCache = make(map[Distances]BranchLengths)
+	var currDistances phylo.AccelDistances
+	var distanceCache = make(map[phylo.AccelDistances]phylo.AccelBranchLengths)
 	var referenceCounter int = 0
 	var reachedEnd bool = false
 	var b1, b3 float64
@@ -120,13 +91,13 @@ func multiFaAcceleration(s Settings) {
 		if records[0].Seq[alignmentCounter] != dna.Gap {                                            //and if we are at a reference position.
 			if pass {
 				if s.UseSnpDistance {
-					reachedEnd = fourWaySnpDistances(records, alignmentCounter, s, &currDistances)
+					reachedEnd = phylo.AccelFourWaySnpDistances(records, alignmentCounter, s.WindowSize, &currDistances)
 				} else {
-					reachedEnd = fourWayMutationDistances(records, alignmentCounter, s, &currDistances)
+					reachedEnd = phylo.AccelFourWayMutationDistances(records, alignmentCounter, s.WindowSize, &currDistances)
 				}
 
 				if _, containedInMap = distanceCache[currDistances]; !containedInMap { //if this tree has not been seen before, calculate branch lengths
-					distanceCache[currDistances] = alternatingLeastSquares(currDistances, s)
+					distanceCache[currDistances] = phylo.BranchLengthsAlternatingLeastSquares(currDistances, s.AllowNegative, s.Verbose, s.ZeroDistanceWeightConstant, s.Epsilon)
 				}
 				//now our distances should be in the cache, and we can do a simple lookup.
 				b1 = distanceCache[currDistances].B1
@@ -150,6 +121,13 @@ func multiFaAcceleration(s Settings) {
 	velBed := fileio.EasyCreate(s.VelOut)
 	accelBed := fileio.EasyCreate(s.AccelOut)
 	initialVelBed := fileio.EasyCreate(s.InitialVelOut)
+	var b1OutBed, b3OutBed *fileio.EasyWriter
+	if s.B1Out != "" {
+		b1OutBed = fileio.EasyCreate(s.B1Out)
+	}
+	if s.B3Out != "" {
+		b3OutBed = fileio.EasyCreate(s.B3Out)
+	}
 
 	//with our normalization parameters calculated, we can normalize the branch lengths for each window and write to file.
 	for i = range branchCacheSlice {
@@ -158,6 +136,12 @@ func multiFaAcceleration(s Settings) {
 		bed.WriteBed(velBed, bed.Bed{Chrom: s.ChromName, ChromStart: branchCacheSlice[i].ChromStart, ChromEnd: branchCacheSlice[i].ChromEnd, Name: fmt.Sprintf("%e", b1Normal), FieldsInitialized: 4})
 		bed.WriteBed(initialVelBed, bed.Bed{Chrom: s.ChromName, ChromStart: branchCacheSlice[i].ChromStart, ChromEnd: branchCacheSlice[i].ChromEnd, Name: fmt.Sprintf("%e", b3Normal), FieldsInitialized: 4})
 		bed.WriteBed(accelBed, bed.Bed{Chrom: s.ChromName, ChromStart: branchCacheSlice[i].ChromStart, ChromEnd: branchCacheSlice[i].ChromEnd, Name: fmt.Sprintf("%e", b1Normal-b3Normal), FieldsInitialized: 4})
+		if s.B1Out != "" {
+			bed.WriteBed(b1OutBed, bed.Bed{Chrom: s.ChromName, ChromStart: branchCacheSlice[i].ChromStart, ChromEnd: branchCacheSlice[i].ChromEnd, Name:fmt.Sprintf("%e", branchCacheSlice[i].B1), FieldsInitialized: 4})
+		}
+		if s.B3Out != "" {
+			bed.WriteBed(b3OutBed, bed.Bed{Chrom: s.ChromName, ChromStart: branchCacheSlice[i].ChromStart, ChromEnd: branchCacheSlice[i].ChromEnd, Name:fmt.Sprintf("%e", branchCacheSlice[i].B3), FieldsInitialized: 4})
+		}
 	}
 
 	err = velBed.Close()
@@ -166,189 +150,17 @@ func multiFaAcceleration(s Settings) {
 	exception.PanicOnErr(err)
 	err = initialVelBed.Close()
 	exception.PanicOnErr(err)
-}
-
-//this helper function calculates the optimal branch lengths for a given set of distances. See readme for a detailed description of this algorithm.
-func alternatingLeastSquares(d Distances, s Settings) BranchLengths {
-	var answer = BranchLengths{1, 1, 1, 1, 1}
-	var Q float64 = calculateQ(d, answer, s)
-	var nextQ float64
-	var currDiff float64 = s.Epsilon + 1 //set currDiff to something larger than epsilon so that we make it into the loop the first time.
-	var sub SubTree
-	var maxIteration, i = 1000, 0
-	var oldAnswer = BranchLengths{1, 1, 1, 1, 1}
-
-	for currDiff > s.Epsilon && i < maxIteration {
-		oldAnswer = answer
-		pruneLeft(d, answer, &sub, s.ZeroDistanceWeightConstant)
-		answer.B1, answer.B2, answer.B3 = optimizeSubtree(&sub, s)
-		pruneRight(d, answer, &sub, s.ZeroDistanceWeightConstant)
-		answer.B4, answer.B5, answer.B3 = optimizeSubtree(&sub, s)
-		nextQ = calculateQ(d, answer, s)
-		//DEBUG: log.Printf("nextQ: %e. currDiff: %e. Here were the branch lengths: %f. %f. %f. %f. %f.", nextQ, currDiff, answer.B1, answer.B2, answer.B3, answer.B4, answer.B5)
-		currDiff = math.Abs(Q - nextQ)
-		if nextQ > Q { //nextQ is higher than Q, which means we got "worse"
-			answer = oldAnswer //we will exit the loop next time, so we want the old answer, which has the lower of the two terminal Q estimates.
-			currDiff = 0
-		}
-		Q = nextQ
-		i++
+	if s.B1Out != "" {
+		err = b1OutBed.Close()
+		exception.PanicOnErr(err)
 	}
-	if i >= maxIteration {
-		log.Fatalf("Failed to converge on a tree with these distances. D01: %f, D02: %f, D03: %f, D12: %f, D13: %f, D23: %f.", d.D01, d.D02, d.D03, d.D12, d.D13, d.D23)
-	}
-	return answer
-}
-
-//a helper function of alternatingLeastSquares. Calculates the optimal branch lengths for the three branches in a subtree.
-func optimizeSubtree(sub *SubTree, s Settings) (float64, float64, float64) {
-	sub.Va = (sub.Dab + sub.Dac - sub.Dbc) / 2.0
-	sub.Vb = (sub.Dab + sub.Dbc - sub.Dac) / 2.0
-	sub.Vc = (sub.Dac + sub.Dbc - sub.Dac) / 2.0
-
-	if s.AllowNegative {
-		return sub.Va, sub.Vb, sub.Vc
-	}
-	if sub.Va < 0 && sub.Vb < 0 && sub.Vc < 0 {
-		if s.Verbose {
-			log.Printf("WARNING: All branches are negative.") //TODO: Should this error out?
-		}
-		sub.Va, sub.Vb, sub.Vc = 0, 0, 0
-	} else if sub.Va < 0 && sub.Vb < 0 {
-		sub.Va = 0
-		sub.Vb = 0
-		sub.Vc = nonNegativeApproximation(sub.Dac, sub.Dbc, sub.Va, sub.Vb, s.ZeroDistanceWeightConstant)
-	} else if sub.Va < 0 && sub.Vc < 0 {
-		sub.Va = 0
-		sub.Vc = 0
-		sub.Vb = nonNegativeApproximation(sub.Dbc, sub.Dab, sub.Vc, sub.Va, s.ZeroDistanceWeightConstant)
-	} else if sub.Vb < 0 && sub.Vc < 0 {
-		sub.Vb = 0
-		sub.Vc = 0
-		sub.Va = nonNegativeApproximation(sub.Dab, sub.Dac, sub.Vb, sub.Vc, s.ZeroDistanceWeightConstant)
-	} else if sub.Va < 0 {
-		sub.Va = 0
-		sub.Vb = nonNegativeApproximation(sub.Dab, sub.Dbc, sub.Va, sub.Vc, s.ZeroDistanceWeightConstant)
-		sub.Vc = nonNegativeApproximation(sub.Dac, sub.Dbc, sub.Va, sub.Vb, s.ZeroDistanceWeightConstant)
-	} else if sub.Vb < 0 {
-		sub.Vb = 0
-		sub.Va = nonNegativeApproximation(sub.Dab, sub.Dac, sub.Vb, sub.Vc, s.ZeroDistanceWeightConstant)
-		sub.Vc = nonNegativeApproximation(sub.Dab, sub.Dbc, sub.Va, sub.Vc, s.ZeroDistanceWeightConstant)
-	} else if sub.Vc < 0 {
-		sub.Vc = 0
-		sub.Va = nonNegativeApproximation(sub.Dab, sub.Dac, sub.Vb, sub.Vc, s.ZeroDistanceWeightConstant)
-		sub.Vb = nonNegativeApproximation(sub.Dab, sub.Dbc, sub.Va, sub.Vc, s.ZeroDistanceWeightConstant)
-	}
-	return sub.Va, sub.Vb, sub.Vc
-}
-
-//If we constrain branch lengths to be nonNegative, we apply this correction when the minimum Q is achieved at negative branch lengths for a subtree.
-func nonNegativeApproximation(d1 float64, d2 float64, v1 float64, v2 float64, ZeroDistanceWeightConstant float64) float64 {
-	if d1 == 0 {
-		if d2 == 0 {
-			return numbers.MaxFloat64(0, (ZeroDistanceWeightConstant*(d1-v1)+ZeroDistanceWeightConstant*(d2-v2))/(2*ZeroDistanceWeightConstant))
-		} else {
-			return numbers.MaxFloat64(0, ZeroDistanceWeightConstant*(d1-v1)+(1.0/math.Pow(d2, 2)*(d2-v2))) / (ZeroDistanceWeightConstant + (1.0 / math.Pow(d2, 2)))
-		}
-	} else if d2 == 0 {
-		return numbers.MaxFloat64(0, (1.0/(math.Pow(d1, 2))*(d1-v1)+ZeroDistanceWeightConstant*(d2-v2))/((1.0/math.Pow(d1, 2))+ZeroDistanceWeightConstant))
-	}
-	return numbers.MaxFloat64(0, (1.0/(math.Pow(d1, 2))*(d1-v1)+(1.0/math.Pow(d2, 2)*(d2-v2)))/((1.0/math.Pow(d1, 2))+(1.0/math.Pow(d2, 2))))
-}
-
-//Reduce the four species tree to the subtree containing species 0, 1, and the ancestor of 2/3.
-func pruneLeft(d Distances, b BranchLengths, sub *SubTree, ZeroDistanceWeightConstant float64) {
-	sub.Dab = d.D01
-	if d.D03 == 0 {
-		if d.D02 == 0 {
-			sub.Dac = (ZeroDistanceWeightConstant*(d.D02-b.B4) + ZeroDistanceWeightConstant*(d.D03-b.B5)) / (2 * ZeroDistanceWeightConstant)
-		} else {
-			sub.Dac = ((1.0/math.Pow(d.D02, 2))*(d.D02-b.B4) + (ZeroDistanceWeightConstant * (d.D03 - b.B5))) / (ZeroDistanceWeightConstant + (1.0 / math.Pow(d.D02, 2)))
-		}
-	} else if d.D02 == 0 {
-		sub.Dac = (ZeroDistanceWeightConstant*(d.D02-b.B4) + ((1.0 / math.Pow(d.D03, 2)) * (d.D03 - b.B5))) / ((1.0 / math.Pow(d.D03, 2)) + ZeroDistanceWeightConstant)
-	} else {
-		sub.Dac = ((1.0/math.Pow(d.D02, 2))*(d.D02-b.B4) + (1.0/math.Pow(d.D03, 2))*(d.D03-b.B5)) / ((1.0 / math.Pow(d.D03, 2)) + (1.0 / math.Pow(d.D02, 2)))
-	}
-
-	if d.D13 == 0 {
-		if d.D12 == 0 {
-			sub.Dbc = (ZeroDistanceWeightConstant*(d.D12-b.B4) + ZeroDistanceWeightConstant*(d.D13-b.B5)) / (2 * ZeroDistanceWeightConstant)
-		} else {
-			sub.Dbc = (ZeroDistanceWeightConstant*(d.D13-b.B5) + (1.0/math.Pow(d.D12, 2))*(d.D12-b.B4)) / ((1.0 / math.Pow(d.D12, 2)) + ZeroDistanceWeightConstant)
-		}
-	} else if d.D12 == 0 {
-		sub.Dbc = (ZeroDistanceWeightConstant*(d.D12-b.B4) + (1.0/math.Pow(d.D13, 2))*(d.D13-b.B5)) / ((1.0 / math.Pow(d.D13, 2)) + ZeroDistanceWeightConstant)
-	} else {
-		sub.Dbc = ((1.0/math.Pow(d.D12, 2))*(d.D12-b.B4) + (1.0/math.Pow(d.D13, 2))*(d.D13-b.B5)) / ((1.0 / math.Pow(d.D13, 2)) + (1.0 / math.Pow(d.D12, 2)))
+	if s.B3Out != "" {
+		err = b3OutBed.Close()
+		exception.PanicOnErr(err)
 	}
 }
 
-//Reduce the four species tree to the subtree containing 2, 3, and the ancestor of 0/1.
-func pruneRight(d Distances, b BranchLengths, sub *SubTree, ZeroDistanceWeightConstant float64) {
-	sub.Dac = d.D23
-
-	if d.D02 == 0 {
-		if d.D12 == 0 {
-			sub.Dac = (ZeroDistanceWeightConstant*(d.D02-b.B1) + ZeroDistanceWeightConstant*(d.D12-b.B2)) / (2.0 * ZeroDistanceWeightConstant)
-		} else {
-			sub.Dac = ((1.0/math.Pow(d.D12, 2))*(d.D12-b.B2) + ZeroDistanceWeightConstant*(d.D02-b.B1)) / ((1.0 / math.Pow(d.D12, 2)) + ZeroDistanceWeightConstant)
-		}
-	} else if d.D12 == 0 {
-		sub.Dac = ((1.0/math.Pow(d.D02, 2))*(d.D02-b.B1) + ZeroDistanceWeightConstant*(d.D12-b.B2)) / ((1.0 / math.Pow(d.D02, 2)) + ZeroDistanceWeightConstant)
-	} else {
-		sub.Dab = ((1.0/math.Pow(d.D02, 2))*(d.D02-b.B1) + (1.0/math.Pow(d.D12, 2))*(d.D12-b.B2)) / ((1.0 / math.Pow(d.D02, 2)) + (1.0 / math.Pow(d.D12, 2)))
-	}
-
-	if d.D03 == 0 {
-		if d.D13 == 0 {
-			sub.Dbc = (ZeroDistanceWeightConstant*(d.D03-b.B1) + ZeroDistanceWeightConstant*(d.D13-b.B2)) / (2.0 * ZeroDistanceWeightConstant)
-		} else {
-			sub.Dbc = ((1.0/math.Pow(d.D13, 2))*(d.D13-b.B2) + ZeroDistanceWeightConstant*(d.D03-b.B1)) / ((1.0 / math.Pow(d.D13, 2)) + ZeroDistanceWeightConstant)
-		}
-	} else if d.D13 == 0 {
-		sub.Dbc = ((1.0/math.Pow(d.D03, 2))*(d.D03-b.B1) + ZeroDistanceWeightConstant*(d.D13-b.B2)) / ((1.0 / math.Pow(d.D03, 2)) + ZeroDistanceWeightConstant)
-	} else {
-		sub.Dbc = ((1.0/math.Pow(d.D03, 2))*(d.D03-b.B1) + (1.0/math.Pow(d.D13, 2))*(d.D13-b.B2)) / ((1.0 / math.Pow(d.D03, 2)) + (1.0 / math.Pow(d.D13, 2)))
-	}
-}
-
-//For a set of distances and corresponding branch lengths, determine the value of Q, the Fitch-Margoliash least squares error.
-func calculateQ(d Distances, b BranchLengths, s Settings) float64 {
-	var sum float64 = 0
-	if d.D01 != 0 { //avoid divide by zero error
-		sum += math.Pow(d.D01-b.B1-b.B2, 2) / math.Pow(d.D01, 2)
-	} else {
-		sum += math.Pow(d.D01-b.B1-b.B2, 2) * s.ZeroDistanceWeightConstant
-	}
-	if d.D02 != 0 {
-		sum += math.Pow(d.D02-b.B1-b.B3-b.B4, 2) / math.Pow(d.D02, 2)
-	} else {
-		sum += math.Pow(d.D02-b.B1-b.B3-b.B4, 2) * s.ZeroDistanceWeightConstant
-	}
-	if d.D03 != 0 {
-		sum += math.Pow(d.D03-b.B1-b.B3-b.B5, 2) / math.Pow(d.D03, 2)
-	} else {
-		sum += math.Pow(d.D03-b.B1-b.B3-b.B5, 2) * s.ZeroDistanceWeightConstant
-	}
-	if d.D12 != 0 {
-		sum += math.Pow(d.D12-b.B2-b.B3-b.B4, 2) / math.Pow(d.D12, 2)
-	} else {
-		sum += math.Pow(d.D12-b.B2-b.B3-b.B4, 2) * s.ZeroDistanceWeightConstant
-	}
-	if d.D13 != 0 {
-		sum += math.Pow(d.D13-b.B2-b.B3-b.B5, 2) / math.Pow(d.D13, 2)
-	} else {
-		sum += math.Pow(d.D13-b.B2-b.B3-b.B5, 2) * s.ZeroDistanceWeightConstant
-	}
-	if d.D23 != 0 {
-		sum += math.Pow(d.D23-b.B4-b.B5, 2) / math.Pow(d.D23, 2)
-	} else {
-		sum += math.Pow(d.D23-b.B4-b.B5, 2) * s.ZeroDistanceWeightConstant
-	}
-	return sum
-}
-
+//this helper function returns true if a sufficient number of positions in the candidate window overlap the search space, as measured by non-zero values in the bitArray.
 //bitArray is on reference coordinates, not alignment coordinates, so the window is simply equal to windowSize.
 func thresholdCheckPasses(s Settings, currCount int, threshold int, bitArray []bool, referenceCounter int) (int, bool) {
 	if s.SearchSpaceBed == "" { //no search space file, no need to look further
@@ -372,85 +184,6 @@ func thresholdCheckPasses(s Settings, currCount int, threshold int, bitArray []b
 	return currCount, currCount >= threshold
 }
 
-//Generate distances from mutation distances, which includes SNPs and INDELs, where each INDEL counts as one mutation regardless of length.
-func fourWayMutationDistances(records []fasta.Fasta, alignmentCounter int, s Settings, D *Distances) bool {
-	//first we clear the values in D.
-	D.D01, D.D02, D.D03, D.D12, D.D13, D.D23 = 0, 0, 0, 0, 0, 0
-	var D01tmp int
-	var reachedEnd bool
-	var alnEnd int
-	D01tmp, reachedEnd, alnEnd = fasta.PairwiseMutationDistanceReferenceWindow(records[0], records[1], alignmentCounter, s.WindowSize)
-	D.D01 = float64(D01tmp)
-	D.D02 = float64(fasta.PairwiseMutationDistanceInRange(records[0], records[2], alignmentCounter, alnEnd))
-	D.D03 = float64(fasta.PairwiseMutationDistanceInRange(records[0], records[3], alignmentCounter, alnEnd))
-	D.D12 = float64(fasta.PairwiseMutationDistanceInRange(records[1], records[2], alignmentCounter, alnEnd))
-	D.D13 = float64(fasta.PairwiseMutationDistanceInRange(records[1], records[3], alignmentCounter, alnEnd))
-	D.D23 = float64(fasta.PairwiseMutationDistanceInRange(records[2], records[3], alignmentCounter, alnEnd))
-	return reachedEnd
-}
-
-//Generate distances from SNP distance, which includes only SNPs.
-func fourWaySnpDistances(records []fasta.Fasta, alignmentCounter int, s Settings, d *Distances) bool {
-	//first we clear the values in d.
-	d.D01, d.D02, d.D03, d.D12, d.D13, d.D23 = 0, 0, 0, 0, 0, 0
-	var baseCount, i int = 0, 0
-	var reachedEnd bool = false
-
-	if len(records) != 4 {
-		log.Fatalf("multiFaAcceleration must take in a four-way multiple alignment.")
-	}
-	for i = alignmentCounter; baseCount < s.WindowSize && i < len(records[0].Seq); i++ {
-		if records[0].Seq[i] != dna.Gap {
-			baseCount++
-		}
-		if isUngappedColumn(records, i) {
-			if records[0].Seq[i] != records[1].Seq[i] {
-				d.D01++
-			}
-			if records[0].Seq[i] != records[2].Seq[i] {
-				d.D02++
-			}
-			if records[0].Seq[i] != records[3].Seq[i] {
-				d.D03++
-			}
-			if records[1].Seq[i] != records[2].Seq[i] {
-				d.D12++
-			}
-			if records[1].Seq[i] != records[3].Seq[i] {
-				d.D13++
-			}
-			if records[2].Seq[i] != records[3].Seq[i] {
-				d.D23++
-			}
-		}
-	}
-	if baseCount != s.WindowSize {
-		reachedEnd = true
-	}
-	return reachedEnd
-}
-
-//a helper function of fourWaySnpDistances, determines if an alignment column is comprised of bases (not gaps) for each species.
-func isUngappedColumn(records []fasta.Fasta, index int) bool {
-	for i := range records {
-		if !isUngappedBase(records[i].Seq[index]) {
-			return false
-		}
-	}
-	return true
-}
-
-//a helper function of isUngappedColumn. True if a dna.Base is a base, not an N, gap, or dot.
-func isUngappedBase(b dna.Base) bool {
-	if b == dna.A || b == dna.T || b == dna.C || b == dna.G {
-		return true
-	}
-	if b == dna.LowerA || b == dna.LowerC || b == dna.LowerG || b == dna.LowerT {
-		return true
-	}
-	return false
-}
-
 func usage() {
 	fmt.Print(
 		"multiFaAcceleration - Performs velocity and acceleration on a four way multiple alignment in multiFa format." +
@@ -472,6 +205,8 @@ func main() {
 	var epsilon *float64 = flag.Float64("epsilon", 1e-8, "Set the error threshold for alternating least squares branch length calculation.")
 	var allowNegative *bool = flag.Bool("allowNegative", false, "Allow the algorithm to evaluate negative branch lengths. This program will constrain the optimal solution to non-negative branch lengths by default.")
 	var zeroDistanceWeightConstant *float64 = flag.Float64("zeroDistanceWeightConstant", 1000, "Set the relative error weight applied to pairs of species with a pairwise distance of zero.")
+	var rawVelBranchLength *string = flag.String("rawVelBranchLength", "", "Set an output file name to return the raw branch length for the branch associated with the velocity score.")
+	var rawInitialVelBranchLength *string = flag.String("rawInitialVelBranchLength", "", "Set an output file name to return the raw branch length for the branch associated with the initial velocity score.")
 
 	flag.Usage = usage
 	log.SetFlags(log.Ldate | log.Ltime | log.Lshortfile)
@@ -502,6 +237,8 @@ func main() {
 		Epsilon:                    *epsilon,
 		AllowNegative:              *allowNegative,
 		ZeroDistanceWeightConstant: *zeroDistanceWeightConstant,
+		B1Out: *rawVelBranchLength,
+		B3Out: *rawInitialVelBranchLength,
 	}
 
 	multiFaAcceleration(s)
